Massive refactor to clock tower
