+++ clocktower/app/tower/[id]/actions/fetchClockDataSA.ts
@@ -1,36 +1,36 @@
 'use server'
-// Fetch Row Data
+// Fetch Clock Data
 import { createServerActionClient } from '@supabase/auth-helpers-nextjs'
 import {
   UUID,
+  ClockType,
   UUIDSchema,
+  ClockSchema,
   ServerActionReturn,
-  TowerRowRow,
-  TowerRowRowSchema,
 } from '@/types/schemas'
 import { Database } from '@/types/supabase'
 import { cookies } from 'next/headers'
 
-export default async function fetchTowerRowData(
-  inputRowId: UUID,
-): Promise<ServerActionReturn<TowerRowRow>> {
+export const fetchClockData = async (
+  inputClockId: UUID,
+): Promise<ServerActionReturn<ClockType>> => {
   try {
     // Test the input with zod, if error, we're checking for errors anyway
-    const rowId = UUIDSchema.parse(inputRowId)
+    const clockId = UUIDSchema.parse(inputClockId)
 
-    // Get the tower row data from the database
+    // Get the clock data from the database
     const supabase = createServerActionClient<Database>({ cookies })
     const { data, error } = await supabase
-      .from('tower_row')
+      .from('clocks')
       .select('*')
-      .eq('id', rowId)
+      .eq('id', clockId)
       .single()
     if (error) throw error
 
-    return { data: TowerRowRowSchema.parse(data) }
+    return { data: ClockSchema.parse(data) }
   } catch (error) {
     return error instanceof Error
       ? { error: error.message }
-      : { error: 'Unknown error from fetchTowerRowData.' }
+      : { error: 'Unknown error from fetchClockData.' }
   }
 }
